Added working FFT Spectrum with fftw.

+++ src/pc-parser/src/eeg-graph.hh
@@ -25,10 +25,10 @@ class EEGGraph
     TPad* m_pad2;
     TGraph* m_graph;
     TH1F* m_fft;
-    unsigned int m_amp_max;
+    double m_amp_max;
 
     void updateGraph(unsigned int points, Double_t* x, Double_t* y);
-    void updateFFT(unsigned int frequencies, unsigned int* values);
+    void updateFFT(unsigned int frequencies, double* values);
     void render();
 
 public:
